Kurakin Matvey. FIIT1. Lab 2. LLVM IR (#58)
The task was to write a pass that combines the fmul and fadd
instructions in llvm.fmuladd

Inside the function, fadd instructions are searched for, in which one of
the operands is the fmul instruction. These instructions are replaced by
llvm.fmuladd and deleted. The fmul instruction is deleted if it is no
longer in use.