Kurakin Matvey. FIIT1. Lab 3. Logic ops (#128)
The task was to write a pass that replaces logical operations.

Inside the function, a search is performed for logical instructions that
are replaced under the conditions:
- If the operation is OR and one of the operands is 0, it is replaced by
a COPY of the other operand.
- If the operation is OR and one of the operands is 1, then it is
replaced by COPY 1.
- If the AND operation and one of the operands is 0, it is replaced by
COPY 0.
- If the AND operation and one of the operands is 1, then it is replaced
by a COPY of the other operand.
- If the OR or AND operation and both operands are equal, then one of
the operands is replaced with COPY.
- All logical operations are replaced by vector analogues.