Kurakin Matvey. FIIT1. Lab 4. Trace loop (#163)
The task was to write a pass that adds a call to the
@trace_loop_iter_begin() and @trace_loop_iter_end() functions at the
beginning and end of the loop.
Inside the module for affine.for, scf.for, scf.while
@trace_loop_iter_begin() is inserted at the beginning of the loop body,
@trace_loop_iter_end() is inserted before the yield operation. Nested
loops of all supported types are processed. The @trace_loop_iter_begin
and @trace_loop_iter_end. Functions are declared as private if they are
not present in the module.